Ambiente pruebas (#32)
* Simv5 busqueda documentos rey004 (#29)

* sdff

* Ajuste buscar por indices Abril 20 2023

* Ajuste Abril 20 2023 Noche

* dads

* sfdsf

* Ajuste indices Abril 23 2023

---------


* Si mv5 calculo evaluacion val tra rey 005 (#30)

* Ajuste hoy Calculo

* x

* Ajuste Abril 17 2023

* Corrige N_Realcion nulo Abril 23 2023

* Ajuste logo Abril 23 2023

---------


* Ajuste Cache de la layout

* Reposiciones flora (#31)

* Se realizan todos los ajustes solicitados en el formulario de reposiciones

* Mejoras

* Se realizan las mejoras solicitadas al reporte de seguimiento a trámites nuevos

* Se realizan ajustes finales

* rr

---------

+++ SIMv5/Areas/ControlVigilancia/Models/Reposicion.cs
@@ -81,10 +81,10 @@
         public string Asunto { get; set; }
 
         [JsonProperty("coordenadaX")]
-        public decimal? CoordenadaX { get; set; }
+        public float? CoordenadaX { get; set; }
 
         [JsonProperty("coordenadaY")]
-        public decimal? CoordenadaY { get; set; }
+        public float? CoordenadaY { get; set; }
 
         [JsonProperty("nroLeniosSolicitados")]
         public decimal? NroLeniosSolicitados { get; set; }
